pyproject.toml
[build-system]
requires = ["poetry>=1.1.0"]
build-backend = "poetry.masonry.api"

[tool.poetry]
name = "geppetto"
version = "0.0.1"
authors = [
    "Lucas Arbues <lucas.arbues@coinfabrik.com>",
    "Ignacio Peña <ignacio.penia@coinfabrik.com>",
    "Federico Patagua <federico.patagua@coinfabrik.com>",
    "Arturo Beccar <arturo.beccar@coinfabrik.com>",
    "José Garcia Crosta <jose.garcia.crosta@coinfabrik.com>",
    "David Weil <david.weil@coinfabrik.com>",
    "Diego Kelyacoubian <diego.kelyacoubian@coinfabrik.com>",
    "Sebastian Wain <sebastian.wain@coifabrik.com>"
]
description = "Geppetto is a Slack bot that integrates with OpenAIs and DALL-E-3 models."
readme = "README.md" 
classifiers = [
    "Development Status :: 3 - Alpha",
        "Intended Audience :: Developers",
        "License :: OSI Approved :: MIT License",
        "Programming Language :: Python :: 3.6",
        "Topic :: Chatbot",
]

[tool.poetry.dependencies]
python = "^3.8"
certifi = "^2023.11.17"
openai = "^1.4.0"
python-dotenv = "^1.0.0"
slack-bolt = "^1.18.1"
slack-sdk = "^3.26.1"
Pillow = "^10.1.0"
google-generativeai = { version = "*", python = ">=3.9,<4.0" }
IPython = "^8.0.0"

[tool.poetry.scripts]
geppetto = "geppetto.main:main"

[project.urls]
Homepage = "https://github.com/CoinFabrik/geppetto"
Issues = "https://github.com/CoinFabrik/geppetto/issues"